api/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
PyMySQL==1.0.2
python-dotenv==0.21.0
